Add directories to store Models and Routes
